requirements.txt
# FastAPI and server
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9

# Audio processing
pyannote.audio==3.1.1
torch>=2.0.0
torchaudio>=2.0.0

# Transcription
faster-whisper==1.0.1

# HTTP client for webhooks
httpx==0.27.0

# Queue and background tasks
redis==5.0.1
rq==1.16.0

# Utilities
python-dotenv==1.0.1
pydantic==2.6.1
pydantic-settings==2.2.1
numpy>=1.24.0

# Audio format handling
soundfile==0.12.1
librosa==0.10.1
